- correct GPUs attribution from get_device_ids
